Refactor EventProcessor and Server Implementation
- Moved the implementation of the EventProcessor class from event_processor.cpp to a new file server.impl.cpp for better organization.
- Updated server.cpp to include the new server.impl.cpp and modified the main function to call main_call.
- Added a finalize method declaration in event_processor.h, removing the previous implementation from the header file.
- Enhanced the ChromeWriter class by cleaning up unnecessary includes and comments in both chrome_writer.cpp and chrome_writer.h.
- Improved logging and error handling in various functions to ensure better traceability and debugging capabilities.
