+++ src/datacrumbs/server/process/event_processor.h
@@ -4,25 +4,27 @@
 #include <bpf/bpf.h>
 #include <bpf/libbpf.h>
 // Generated Headers
+#include <datacrumbs/common/logging.h>
 #include <datacrumbs/datacrumbs_config.h>
-// other headers
+// Internal Headers
 #include <datacrumbs/common/configuration_manager.h>
-#include <datacrumbs/common/constants.h>
 #include <datacrumbs/common/data_structures.h>
-#include <datacrumbs/common/logging.h>  // Logging header
-#include <datacrumbs/common/singleton.h>
-#include <datacrumbs/common/typedefs.h>
 #include <datacrumbs/common/utils.h>
+#include <datacrumbs/server/bpf/compat/map.h>
 #include <datacrumbs/server/bpf/shared.h>
 #include <datacrumbs/server/process/writer/chrome_writer.h>
+
 // std headers
 #include <errno.h>
+#include <fcntl.h>
 #include <grp.h>
 #include <json-c/json.h>
+#include <limits.h>
 #include <pwd.h>
 #include <stdio.h>
 #include <sys/stat.h>
 #include <sys/types.h>
+#include <sys/utsname.h>
 #include <unistd.h>
 
 #include <atomic>
@@ -53,13 +55,7 @@ class EventProcessor {
     return 0;
   }
 
-  int finalize() {
-    if (writer_) {
-      writer_->finalize();
-    }
-    DC_LOG_DEBUG("EventProcessor finalized");
-    return 0;
-  }
+  int finalize();
 
  public:
   std::shared_ptr<ConfigurationManager> configManager_;
