+++ src/datacrumbs/server/process/server.cpp
@@ -1,3 +1,7 @@
 #include <datacrumbs/server/bpf/datacrumbs.skel.h>
 
-#include <datacrumbs/server/process/event_processor.cpp>
+#include <datacrumbs/server/process/server.impl.cpp>
+
+int main(int argc, char** argv) {
+  return main_call(argc, argv);
+}
